An email conversation with Chuck Fresno turned up several forms of symmetry which were not implemented in Solo. Now they are.
In the process I've completely retired symmetry_limit() on the
grounds that some of the new symmetries do not have a rectangular
base region; instead I determine the base region by going through
the grid and finding every square which is not transformed into a
lexicographically lower square by any symmetry operation. This means
that adding new symmetries is now _only_ a matter of encoding the
actual transformation rules.


git-svn-id: svn://svn.tartarus.org/sgt/puzzles@5965 cda61777-01e9-0310-a592-d414129be87e