Now that we're highlighting the currently selected preset in the Type menu, it looks faintly silly that Fifteen doesn't have any presets other than Custom: you open a Fifteen window in its default state, and the Type menu appears to be telling you it has a custom size! Fixed by adding a preset for the default parameters.
I'd quite like to fix this properly by revamping the presets
mechanism in a way that _enforces_ that there must always be a
preset which matches the default parameters, but that's more fiddly
than it sounds. For the moment, this change fixes the only
externally visible infelicity in the current game set.


git-svn-id: svn://svn.tartarus.org/sgt/puzzles@7983 cda61777-01e9-0310-a592-d414129be87e